Put TeamMember class in its own file
