+++ team_member.py
@@ -1,0 +1,33 @@
+from datetime import datetime
+from typing import List
+from pytz import timezone
+
+class TeamMember:
+    def __init__(self, csv_row: List):
+        self.name = csv_row[0]
+        self.timezone = csv_row[1]
+        self.time = _get_current_formatted_time(self.timezone)
+        self.city = csv_row[2].strip()
+        self._geo_location = None
+
+    @property
+    def _location(self):
+        # This import is slow. Only import it if you need it.
+        from geopy.geocoders import Nominatim
+        if self._geo_location is None:
+            geolocator = Nominatim(user_agent="teamTime")
+            self._geo_location = geolocator.geocode(self.city)
+
+        return self._geo_location
+
+    @property
+    def latitude(self):
+        return self._location.latitude
+
+    @property
+    def longitude(self):
+        return self._location.longitude
+
+
+def _get_current_formatted_time(staff_zone, format_="%Y-%m-%d %H:%M"):
+    return datetime.now(timezone(staff_zone)).strftime(format_)
